rename WelcomeActions to Welcome
